requirements.txt
# Astro Runtime includes the following pre-installed provider packages: 
# https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos>=1.3.0
# Core data processing
pandas>=2.0.0
duckdb>=0.9.0
numpy>=1.24.0

# dbt for transformations (when dbt project is added)
dbt-core>=1.6.0
dbt-duckdb>=1.6.0

# Additional data quality and monitoring
great-expectations>=0.18.0  # For advanced data quality checks

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Core Streamlit and data processing
streamlit>=1.28.0

# Visualization
plotly>=5.17.0

# Utilities
pathlib2>=2.3.7
python-dateutil>=2.8.2
pytz>=2023.3

# Election calendar seeding dependencies
requests>=2.31.0  # For MIT Election Lab and Google Civic API
python-dotenv>=1.0.0  # For optional .env file loading